Make PointStoreHistoryCleaner for clearing data
